==> Part 3.3 - Polarity Analysis with TextBlob - Advanced Clustering.ipynb <==
# %store -r textblob_df1
# X = textblob_df1.iloc[:, -2:].values

# from sklearn.cluster import KMeans
# kmeans = KMeans(n_clusters = 9, init = 'k-means++', random_state = 0)
# y_kmeans = kmeans.fit_predict(X)

# import matplotlib.pyplot as plt
# plt.scatter(X[y_kmeans == 0, 0], X[y_kmeans == 0, 1], s = 20, c = 'red', label = 'C1')
# plt.scatter(X[y_kmeans == 1, 0], X[y_kmeans == 1, 1], s = 20, c = 'blue', label = 'C2')
# plt.scatter(X[y_kmeans == 2, 0], X[y_kmeans == 2, 1], s = 20, c = 'green', label = 'C3')
# plt.scatter(X[y_kmeans == 3, 0], X[y_kmeans == 3, 1], s = 20, c = 'yellow', label = 'C4')
# plt.scatter(X[y_kmeans == 4, 0], X[y_kmeans == 4, 1], s = 20, c = 'orange', label = 'C5')
# plt.scatter(X[y_kmeans == 5, 0], X[y_kmeans == 5, 1], s = 20, c = 'cyan', label = 'C6')
# plt.scatter(X[y_kmeans == 6, 0], X[y_kmeans == 6, 1], s = 20, c = 'magenta', label = 'C7')
# plt.scatter(X[y_kmeans == 7, 0], X[y_kmeans == 7, 1], s = 20, c = 'black', label = 'C8')
# plt.scatter(X[y_kmeans == 8, 0], X[y_kmeans == 8, 1], s = 20, c = 'purple', label = 'C9')

# From these 9 cluster, I can pick the 3 top right corner clusters to carry positive sentiment, 
# the 3 lower left clusters to carry negative sentiment, 
# and the rest (the 3 cluster along the diagonal with a negative slope) to carry neutral sentiment.

# hc3 = AgglomerativeClustering(n_clusters = 9, affinity = 'euclidean', linkage = 'ward')
# y_hc3 = hc3.fit_predict(X)

# plt.scatter(X[y_hc3 == 0, 0], X[y_hc3 == 0, 1], s = 20, c = 'red', label = 'C1')
# plt.scatter(X[y_hc3 == 1, 0], X[y_hc3 == 1, 1], s = 20, c = 'blue', label = 'C2')
# plt.scatter(X[y_hc3 == 2, 0], X[y_hc3 == 2, 1], s = 20, c = 'green', label = 'C3')
# plt.scatter(X[y_hc3 == 3, 0], X[y_hc3 == 3, 1], s = 20, c = 'yellow', label = 'C4')
# plt.scatter(X[y_hc3 == 4, 0], X[y_hc3 == 4, 1], s = 20, c = 'orange', label = 'C5')
# plt.scatter(X[y_hc3 == 5, 0], X[y_hc3 == 5, 1], s = 20, c = 'cyan', label = 'C6')
# plt.scatter(X[y_hc3 == 6, 0], X[y_hc3 == 6, 1], s = 20, c = 'magenta', label = 'C7')
# plt.scatter(X[y_hc3 == 7, 0], X[y_hc3 == 7, 1], s = 20, c = 'black', label = 'C8')
# plt.scatter(X[y_hc3 == 8, 0], X[y_hc3 == 8, 1], s = 20, c = 'purple', label = 'C9')
