==> Part 5.2 - NLP Sentiment Prediction Using Deep Learning - CNN.ipynb <==
# Sentiment Prediction Using Deep Learning - Convolutional Neural Network

## Importing data

import pandas as pd

df = pd.read_csv("./data/dataset.csv", header = None, encoding='latin-1', names=["Sentiment", "Headlines"])
df['Sentiment'] = df['Sentiment'].replace("negative",0).replace("neutral",1).replace("positive",2)
df

df.Sentiment.value_counts()

import sys
sys.path.insert(0, './lib')
from sentiment_module import *

df['Splitted'] = df['Headlines'].apply(lambda x: cleaning_data(x))

df.head(5)

## Splitting data

from sklearn.model_selection import train_test_split
df_train, df_test = train_test_split(df, test_size=0.2, random_state=0)

training_words = [word for tokens in df_train['Splitted'] for word in tokens]

training_words[0:10]

test_words = [word for tokens in df_test['Splitted'] for word in tokens]

test_words[0:10]

## Loading Google News Word2Vec model 

from gensim import models
word2vec_path = './data/GoogleNews-vectors-negative300.bin.gz'
word2vec = models.KeyedVectors.load_word2vec_format(word2vec_path, binary=True)

## Tokenizing & Pad Sequencing

from keras.preprocessing.text import Tokenizer
from keras.preprocessing.sequence import pad_sequences

tokenizer = Tokenizer(num_words=len(training_words), char_level=False)
tokenizer.fit_on_texts(df_train['Headlines'].tolist())
train_sequences = tokenizer.texts_to_sequences(df_train['Headlines'].tolist())
train_word_index = tokenizer.word_index
max_seq_len = max([len(x) for x in train_sequences])

# Pad Sequence
train_data = pad_sequences(train_sequences, maxlen=max_seq_len)

train_data.shape

test_sequences = tokenizer.texts_to_sequences(df_test['Headlines'].tolist())

# Pad Sequence
test_data = pad_sequences(test_sequences, maxlen=max_seq_len)

## Embedding

import numpy as np

train_embeddings = np.zeros((len(train_word_index)+1, 300))
for word,index in train_word_index.items():
    train_embeddings[index,:] = word2vec[word] if word in word2vec else np.random.rand(300)

train_embeddings.shape

## CNN Modeling

import tensorflow as tf
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Dense, Dropout, Conv1D, GlobalMaxPooling1D
from keras.layers.embeddings import Embedding

from keras.utils import np_utils
X_train = train_data
# y_train = df_train['Sentiment']
X_test = test_data
# y_test = df_test['Sentiment']
y_train = np_utils.to_categorical(df_train['Sentiment'], num_classes=3)
y_test = np_utils.to_categorical(df_test['Sentiment'], num_classes=3)

model = Sequential()

model.add(Embedding(input_dim=len(train_word_index)+1, 
                    output_dim=300, 
                    weights=[train_embeddings], 
                    input_length=max_seq_len, 
                    trainable=False))
model.add(Conv1D(filters = 200, kernel_size = 3, padding='valid', activation = 'relu'))
model.add(GlobalMaxPooling1D())
model.add(Dense(128, activation='relu'))
model.add(Dense(3, activation='softmax'))
model.compile(loss='categorical_crossentropy', optimizer='adam', metrics=['accuracy'])
model.fit(X_train, y_train, epochs=10, batch_size=32)

print(X_train.shape)
print(y_train.shape)

model.evaluate(x=X_test, y=y_test, batch_size=None, verbose=1, sample_weight=None)

# Predicting

## CNBC data

### Headlines

%store -r df1
hl1_sequences = tokenizer.texts_to_sequences(df1['Headlines'].tolist())
# Pad Sequence
hl1_data = pad_sequences(hl1_sequences, maxlen=max_seq_len)
y_pred_hl1 = model.predict(hl1_data)

y_pred_hl1[0:10]

from sentiment_module import cluster_extraction
hl_sentiment = cluster_extraction(y_pred_hl1)

hl_sentiment[0:10]

### Description

ds1_sequences = tokenizer.texts_to_sequences(df1['Description'].tolist())
# Pad Sequence
ds1_data = pad_sequences(ds1_sequences, maxlen=max_seq_len)
y_pred_ds1 = model.predict(ds1_data)

y_pred_ds1[0:10]

ds_sentiment = cluster_extraction(y_pred_ds1)
ds_sentiment[0:10]

### Combine

cnn_c_sentiment = combine_sentiments(hl_sentiment, ds_sentiment)
cnn_c_sentiment[0:10]

# storing data for the result dataframe
%store -r final_df1
final_df1['cnn_sentiment'] = cnn_c_sentiment
%store final_df1

## Reuters data

### Headlines

%store -r df2
hl2_sequences = tokenizer.texts_to_sequences(df2['Headlines'].tolist())
# Pad Sequence
hl2_data = pad_sequences(hl2_sequences, maxlen=max_seq_len)
y_pred_hl2 = model.predict(hl2_data)

y_pred_hl2[0:10]

from sentiment_module import cluster_extraction
hl_sentiment = cluster_extraction(y_pred_hl2)

hl_sentiment[0:10]

### Description

ds2_sequences = tokenizer.texts_to_sequences(df2['Description'].tolist())
# Pad Sequence
ds2_data = pad_sequences(ds2_sequences, maxlen=max_seq_len)
y_pred_ds2 = model.predict(ds2_data)

y_pred_ds2[0:10]

ds_sentiment = cluster_extraction(y_pred_ds2)
ds_sentiment[0:10]

### Combine

cnn_r_sentiment = combine_sentiments(hl_sentiment, ds_sentiment)
cnn_r_sentiment[0:10]

# storing data for the result dataframe
%store -r final_df2
final_df2['cnn_sentiment'] = cnn_r_sentiment
%store final_df2

## The Guardian data

### Headlines

%store -r df3
hl3_sequences = tokenizer.texts_to_sequences(df3['Headlines'].tolist())
# Pad Sequence
hl3_data = pad_sequences(hl3_sequences, maxlen=max_seq_len)
y_pred_hl3 = model.predict(hl3_data)

y_pred_hl3[0:10]

from sentiment_module import cluster_extraction
hl_sentiment = cluster_extraction(y_pred_hl3)

hl_sentiment[0:10]

### Combine

cnn_g_sentiment = hl_sentiment
cnn_g_sentiment[0:10]

# storing data for the result dataframe
%store -r final_df3
final_df3['cnn_sentiment'] = cnn_g_sentiment
%store final_df3

final_df1.to_csv("./output/cnbc.csv", index = False)
final_df2.to_csv("./output/reuters.csv", index = False)
final_df3.to_csv("./output/guardian.csv", index = False)
